# internal/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "immerle.DeviceDTO": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "expiresAt": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "lastIp": {
                        "type": "string"
                    },
                    "lastSeen": {
                        "type": "string"
                    },
                    "name": {
                        "example": "Pixel 8",
                        "type": "string"
                    },
                    "userAgent": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "immerle.DevicesResponse": {
                "properties": {
                    "devices": {
                        "items": {
                            "$ref": "#/components/schemas/immerle.DeviceDTO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "immerle.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "already_initialized",
                        "type": "string"
                    },
                    "ok": {
                        "example": false,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.LoginResponse": {
                "properties": {
                    "device": {
                        "$ref": "#/components/schemas/immerle.DeviceDTO"
                    },
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    },
                    "token": {
                        "example": "eyJhbGciOiJIUzI1Ni␣...",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "immerle.OKResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "contact": {
            "name": "immerle"
        },
        "description": "Native immerle endpoints (first-run setup, capability\ndiscovery, friends, activity feed, collaborative playlists and\nsynchronized Jam sessions) that complement the Subsonic /\nOpenSubsonic API served under /rest/.\n\nAuthenticated endpoints reuse Subsonic credentials passed as\nquery parameters: u (username) plus either p (password) or the\ntoken pair t (md5(password+salt)) and s (salt), and c (client\nname). Setup and capability endpoints are unauthenticated.",
        "license": {
            "name": "See repository"
        },
        "title": "immerle extension API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticates with username + password (or Subsonic token auth) and returns a device-session JWT carrying a unique id (jti). Use it as \"Authorization: Bearer \u003cjwt\u003e\". The session is tracked in the devices registry and can be revoked.",
                "parameters": [
                    {
                        "description": "Username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic token md5(password+salt)",
                        "in": "query",
                        "name": "t",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic salt",
                        "in": "query",
                        "name": "s",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client/device name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Device label (defaults to c)",
                        "in": "query",
                        "name": "device",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.LoginResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Log in a device (issue a JWT)",
                "tags": [
                    "devices"
                ]
            }
        },
...
        "/devices": {
            "get": {
                "description": "Lists the caller's active device sessions (one per issued JWT), with last-seen time, IP and user agent.",
                "parameters": [
                    {
                        "description": "Username (or Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.DevicesResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List devices",
                "tags": [
                    "devices"
                ]
            }
        },
        "/devices/revoke": {
            "post": {
                "description": "Revokes a device session by id — the associated JWT can no longer authenticate.",
                "parameters": [
                    {
                        "description": "Username (or Bearer token)",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Device id (jti) to revoke",
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.OKResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Revoke a device",
                "tags": [
                    "devices"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}